\name{fsemipar.internal}
\alias{fsemipar.internal}
\title{
Package fsemipar internal functions
}
\description{
The package includes the following internal functions, based on the code by F. Ferraty, which is available on his website at \url{https://www.math.univ-toulouse.fr/~ferraty/SOFTWARES/NPFDA/index.html}. 
}

\details{
\itemize{
\item \code{approx.spline.deriv}
\item\code{Bspline.ini}
\item\code{fnp.kernel.fit}
\item\code{fnp.kernel.fit.test}
\item\code{fnp.kernel.test}
\item\code{fnp.kNN.fit}
\item\code{fnp.kNN.fit.test}
\item\code{fnp.kNN.fit.test.loc}
\item\code{fnp.kNN.GCV}
\item\code{fnp.kNN.test}
\item\code{fsim.kernel.fit.fixedtheta}
\item\code{fsim.kNN.fit.fixedtheta}
\item\code{fun.kernel}
\item\code{fun.kernel.fixedtheta}
\item\code{fun.kNN}
\item\code{fun.kNN.fixedtheta}
\item\code{funopare.kNN}
\item\code{H.fnp.kernel}
\item\code{H.fnp.kNN}
\item\code{H.fsim.kernel}
\item\code{H.fsim.kNN}
\item\code{interp.spline.deriv}
\item\code{quad}
\item\code{semimetric.deriv}
\item\code{semimetric.interv}
\item\code{semimetric.pca}
\item\code{sfplsim.kernel.fit.fixedtheta}
\item\code{sfplsim.kNN.fit.fixedtheta}
\item\code{Splinemlf}
\item\code{symsolve}
}

}



